pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "norun"
version = "0.1.0"
description = "NORUN - GUI+CLI Windows app runner (Wine + umu-run)"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "Hasib Al Tahsin" }]
dependencies = [
  "typer>=0.12",
  "rich>=13.7",
  "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/HasibAlTahsin/norun"
Source = "https://github.com/HasibAlTahsin/norun"
Issues = "https://github.com/HasibAlTahsin/norun/issues"

[project.scripts]
norun = "norun.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/norun"]

[tool.hatch.build.targets.sdist]
include = [
  "src/norun/**",
  "README.md",
  "LICENSE",
  "MANIFEST.in",
  "packaging/**",
]

